[Devops] Move to a 'matrix' pipeline to improve the device tests feedback
Our bots spend most of the time compiling and deploying rather than
executing tests. This means that a bot with an iOS device is most of the
time not really running tests, but deploying them. With the old pipeline
the average runtime to get feedback for all the device tests is around 6
to 7 hours.

In order to improve the speed in which we get results the new pipeline
spawns 5 different testing jobs and a last 6th job that is a fan in to
report the completion to the monitoring person.

The tests have been divided in the following way:

1. Xamarin tests, wich are all those tests that are not bcl tests or
monotouch.
2. Monotouch tests, since it is a large set of tests.
3. All xUnit based BCL tests.
4. All NUnit based BCL tests.
5. mscorlib tests (all 3 splitted assemblies).

In average the jobs individually take the following times:

1. Xamarin tests: 1h 30m
2. Monotouch tests: 45/50m
3. xUnit based BCL: 1h 30m
4. NUnit based BCL: 45 m
5. mscorlib 1h 30m

In a perfect world with infinite bots, we can get device test results in
under 2 hours, but in reality we might be getting around the 3 hours,
which is half of the time atm.

In order to make things easier for the future different templates in yml
have been created under tools/devops/templates

- tools/devops/templates/device-tests.yml
  Basic template that contains all the required steps needed to run the
  tests on device. This template takes two env vars. One which will
  contains the extra labels to pass to xharness and the context to be
  used to set the device tests status.

- tools/devops/templates/job-matrix.yml
  Main template that allows to create pipelines with all the 5 jobs
  targeting different devices. The template takes a numer of parameters
  to set the targetd device and the capabilities of the agents.

- tools/devops/templates/publish-results.yml
  Small helper template for posting messages.

Once the 5 jobs are completed, a 6th job is kicked that does not need an
iOS/tvOS device (we have more agents without capabilities). The jobs
grabs the artifacts from the previous jos, unzipts them and performs the
publication of the test results to vsts. Once completed it reports the
result in a status to github. This job will execute no matter what were
the results of the previous jobs (notice the always() condition).