[DevOps] Add in the demands a check for the Mac OS X version.
We have situations in which bots have old OSs (Maverik etc..) Because
the deamnd of the job does not check the OS version, the agent picks up
a job and tries to execute it. Once picked up, provisionator will fail
since Xcode requires a newer version of the OS.

This change ensures that the job is not picked up by an old bot meaning
that the tests wont fail. This way we make sure we do not have errors in
the device tests due to bot configurations.